requirements.txt
azure-common==1.1.28
azure-core==1.22.1
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
et-xmlfile==1.1.0
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
numpy==1.22.2
oauthlib==3.2.0
openpyxl==3.0.9
pandas==1.4.1
portalocker==2.3.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.8
